+++ src/main/java/com/lab/app/service/AccidentService.java
@@ -4,11 +4,11 @@ import com.lab.app.dto.AccidentDto;
 
 public interface AccidentService {
 
-    AccidentDto getAccident(int accidentId);
+    AccidentDto getAccident(Long accidentId);
 
     AccidentDto createAccident(AccidentDto accidentDto);
 
-    AccidentDto updateAccident(AccidentDto accidentDto, int accidentId);
+    AccidentDto updateAccident(AccidentDto accidentDto, Long accidentId);
 
-    void deleteAccident(int accidentId);
+    void deleteAccident(Long accidentId);
 }
